requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
Flask-Login==0.6.3
Flask-Caching==2.3.0
flask-wtf==1.2.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.32.3
sendgrid==6.11.0
email-validator==2.2.0
Pillow==10.4.0
qrcode[pil]==7.4.2
beautifulsoup4==4.13.5
numpy==2.3.3
openai==1.107.1
openpyxl==3.1.5
pandas==2.3.2
python-telegram-bot==20.7
reportlab
unidecode
botasaurus
fake-useragent
playwright
schedule
selenium
trafilatura
undetected-chromedriver
urllib3
telegram
weasyprint
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-dance
flask-login
flask-sqlalchemy
gunicorn
numpy
oauthlib
openai
openpyxl
pandas
psycopg2-binary
pyjwt
python-telegram-bot
requests
sendgrid
sqlalchemy
telegram
werkzeug
rarfile
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-dance
flask-login
flask-sqlalchemy
gunicorn
numpy
oauthlib
openai
openpyxl
pandas
psycopg2-binary
pyjwt
python-telegram-bot
requests
sendgrid
sqlalchemy
telegram
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
beautifulsoup4==4.13.5
botasaurus
email-validator==2.2.0
fake-useragent
Flask==2.3.3
Flask-Caching==2.3.0
flask-dance
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-wtf==1.2.2
gunicorn==21.2.0
numpy==2.3.3
oauthlib
openai==1.107.1
openpyxl==3.1.5
pandas==2.3.2
Pillow==10.4.0
playwright
psycopg2-binary==2.9.9
pyjwt
python-telegram-bot==20.7
qrcode[pil]==7.4.2
rarfile
reportlab
requests==2.32.3
schedule
selenium
sendgrid==6.11.0
SQLAlchemy==2.0.32
telegram
trafilatura
undetected-chromedriver
unidecode
urllib3
weasyprint
Werkzeug==2.3.7
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
beautifulsoup4==4.13.5
botasaurus
email-validator==2.2.0
fake-useragent
Flask==2.3.3
Flask-Caching==2.3.0
flask-dance
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-wtf==1.2.2
gunicorn==21.2.0
numpy==2.3.3
oauthlib
openai==1.107.1
openpyxl==3.1.5
pandas==2.3.2
Pillow==10.4.0
playwright
psycopg2-binary==2.9.9
pyjwt
python-telegram-bot==20.7
qrcode[pil]==7.4.2
rarfile
reportlab
requests==2.32.3
schedule
selenium
sendgrid==6.11.0
SQLAlchemy==2.0.32
telegram
trafilatura
undetected-chromedriver
unidecode
urllib3
weasyprint
Werkzeug==2.3.7
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
beautifulsoup4==4.13.5
email-validator==2.2.0
fake-useragent
Flask==2.3.3
Flask-Caching==2.3.0
flask-dance
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-wtf==1.2.2
gunicorn==21.2.0
numpy==2.3.3
oauthlib
openai==1.107.1
openpyxl==3.1.5
pandas==2.3.2
Pillow==10.4.0
psycopg2-binary==2.9.9
pyjwt
python-telegram-bot==20.7
qrcode[pil]==7.4.2
rarfile
reportlab
requests==2.32.3
sendgrid==6.11.0
SQLAlchemy==2.0.32
telegram
unidecode
urllib3
weasyprint
Werkzeug==2.3.7
dadata
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
pytest
pytest-mock
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
Flask-Limiter
Flask-Talisman
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email-validator
flask
flask-dance
flask-login
flask-sqlalchemy
gunicorn
numpy
oauthlib
openai
openpyxl
pandas
psycopg2-binary
pyjwt
python-telegram-bot
requests
sendgrid
sqlalchemy
telegram
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
flask-login
apscheduler
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
apscheduler
beautifulsoup4
botasaurus
dadata
email-validator
fake-useragent
Flask
Flask-Caching
flask-dance
Flask-Limiter
Flask-Login
Flask-SQLAlchemy
Flask-Talisman
flask-wtf
gunicorn
numpy
oauthlib
openai
openpyxl
pandas
Pillow
playwright
psycopg2-binary
pyjwt
pytest
pytest-mock
python-telegram-bot
qrcode
rarfile
reportlab
requests
schedule
selenium
sendgrid
SQLAlchemy
telegram
trafilatura
undetected-chromedriver
unidecode
urllib3
weasyprint
Werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
apscheduler
beautifulsoup4
botasaurus
dadata
fake-useragent
Flask-Caching
Flask-Limiter
Flask-Talisman
flask-wtf
Pillow
playwright
pytest
pytest-mock
qrcode
rarfile
reportlab
schedule
selenium
trafilatura
undetected-chromedriver
unidecode
urllib3
weasyprint
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
apscheduler
beautifulsoup4
botasaurus
dadata
email-validator
fake-useragent
Flask
Flask-Caching
flask-dance
Flask-Limiter
Flask-Login
Flask-SQLAlchemy
Flask-Talisman
flask-wtf
gunicorn
numpy
oauthlib
openai
openpyxl
pandas
Pillow
playwright
psycopg2-binary
pyjwt
pytest
pytest-mock
python-telegram-bot
qrcode
rarfile
reportlab
requests
schedule
selenium
sendgrid
SQLAlchemy
telegram
trafilatura
undetected-chromedriver
unidecode
urllib3
weasyprint
flask-caching
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
